Rename typeOfAp page => ApType page
This matches what we have in the code under test, so is easier to find
